Path rename so all workspace files are in the same place
